+++ src/backend/reactions/reaction.cpp
@@ -6,7 +6,7 @@
 #include "pointers.hpp"
 #include "script.hpp"
 #include "services/player_database/player_database_service.hpp"
-#include "util/notify.hpp"
+#include "util/chat.hpp"
 
 namespace big
 {
@@ -62,15 +62,11 @@ namespace big
 
 		if (announce_in_chat)
 		{
-			g_fiber_pool->queue_job([player, this] {
-				auto chat = std::format("{} {}", g.session.chat_output_prefix, std::vformat(g_translation_service.get_translation(m_announce_message), std::make_format_args(player->get_name())));
+			auto msg = std::format("{} {}",
+			    g.session.chat_output_prefix,
+			    std::vformat(g_translation_service.get_translation(m_announce_message), std::make_format_args(player->get_name())));
 
-				if (g_hooking->get_original<hooks::send_chat_message>()(*g_pointers->m_gta.m_send_chat_ptr,
-				        g_player_service->get_self()->get_net_data(),
-				        chat.data(),
-				        is_team_only))
-					notify::draw_chat(chat.c_str(), g_player_service->get_self()->get_name(), is_team_only);
-			});
+			chat::send_message(msg);
 		}
 
 		if (notify)
